requirements.txt
tensorflow
matplotlib
numpy
keras
fastapi
uvicorn
pillow